perf(sessions): reuse import identity query compilation
